Fix tests

+++ tests/phpunit/tests/class-cache.php
@@ -47,7 +47,9 @@ class Test_Cache extends TestCase {
 	 * Dont save cache
 	 */
 	public function test_create_dir() {
-		unlink( ROOT_DIR . 'cache/classmap.php' );
+		if ( file_exists( ROOT_DIR . 'cache/classmap.php' ) ) {
+			unlink( ROOT_DIR . 'cache/classmap.php' );
+		}
 		rmdir( ROOT_DIR . 'cache/' );
 		$cache = new Cache();
 		$cache->update( '\Prefix\Autoload_Success_1', __DIR__ . '/../classes/path-1/prefix/class-autoload-success-1.php' );
